requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
colorlog==4.1.0
configparser==4.0.2
docopt==0.6.2
grafana-api==1.0.3
graphql-core==3.1.3
graphviz==0.16
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
jsonformatter==0.3.1
packaging==20.8
pip-tools==5.5.0
pipdeptree==2.0.0
pipreqs==0.4.10
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
sgqlc==12.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
yarg==0.1.9
zipp==3.4.0
